requirements.txt
appdirs==1.4.4
appscript==1.2.0
atomicwrites==1.4.0
attrs==22.1.0
backports.weakref==1.0.post1
beautifulsoup4==4.11.1
bkcharts==0.2
black==19.10b0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
contourpy==1.0.5
cssselect==1.1.0
cycler==0.11.0
cytoolz==0.11.0
daal==2021.5.0
daal4py==2021.5.0
Deprecated==1.2.13
dpcpp-cpp-rt==2022.1.0
et-xmlfile==1.1.0
fonttools==4.25.0
greenlet==1.1.3
HeapDict==1.0.1
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
intel-cmplr-lic-rt==2022.1.0
intel-openmp==2022.1.0
jedi==0.18.1
kiwisolver==1.4.4
llvmlite==0.38.0
lxml==4.9.1
matplotlib==3.6.0
mccabe==0.6.1
mkl==2022.1.0
mkl-fft==1.3.1
mkl-service==2.4.0
MLB-StatsAPI==1.5.1
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
Pillow==9.0.1
pluggy==1.0.0
protobuf==3.19.1
psutil==5.9.2
py==1.11.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybaseball==2.2.1
pycosat==0.6.3
pycparser==2.21
pycurl==7.44.1
PyDispatcher==2.0.5
PyGithub==1.55
PyJWT==2.5.0
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.1
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
pytz==2021.3
PyYAML==6.0
queuelib==1.5.0
regex==2022.9.13
requests==2.28.1
scipy==1.9.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
tabulate==0.8.9
tbb==2021.6.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
typed-ast==1.5.4
ujson==5.5.0
urllib3==1.26.12
webencodings==0.5.1
wrapt==1.14.1
xlwings==0.24.9
zict==2.0.0
